refactor: 拆分 PostRepository 為三職責特定 Repository
- 新增 PostBaseRepository 抽象類別萃取共用輔助方法與常數
- 新增 HasTransactionSupport trait 封裝交易邏輯
- 新增 PostCacheInvalidator 服務集中管理快取失效
- 新增 PostCrudRepository（CRUD + 標籤操作，~550 行）
- 新增 PostSearchRepository（搜尋方法，~50 行）
- 新增 PostAnalyticsRepository（來源分析查詢，~160 行）
- 瘦身 PostRepository 為 Facade 委派模式（~60 行）
- 補齊 PostRepositoryInterface 缺少的 3 個方法宣告
- 更新 DI 容器與 AttachmentService 型別依賴
